requirements.txt
matplotlib==3.5.3
numpy==1.24.3
pandas==2.0.3
Pillow==9.1.1
Pillow==11.0.0
scikit_learn==1.5.2
spacy==3.8.2
torch==1.8.1
torchvision==0.9.1
